+++ tests/TestCase.php
@@ -1,0 +1,36 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Ultraviolettes\FilamentAudio\Tests;
+
+use Filament\FilamentServiceProvider;
+use Filament\Forms\FormsServiceProvider;
+use Filament\Infolists\InfolistsServiceProvider;
+use Filament\Support\SupportServiceProvider;
+use Filament\Tables\TablesServiceProvider;
+use Livewire\LivewireServiceProvider;
+use Orchestra\Testbench\TestCase as Orchestra;
+use Ultraviolettes\FilamentAudio\AudioServiceProvider;
+
+abstract class TestCase extends Orchestra
+{
+    protected function getPackageProviders($app): array
+    {
+        return [
+            LivewireServiceProvider::class,
+            SupportServiceProvider::class,
+            FormsServiceProvider::class,
+            TablesServiceProvider::class,
+            InfolistsServiceProvider::class,
+            FilamentServiceProvider::class,
+            AudioServiceProvider::class,
+        ];
+    }
+
+    protected function getEnvironmentSetUp($app): void
+    {
+        $app['config']->set('database.default', 'testing');
+        $app['config']->set('view.paths', [__DIR__.'/../resources/views']);
+    }
+}